Add 006_add_project_id_to_equipment_assets.rb
to fix an internal server error when performing a search from any
project. Without this migration the `equipment_assets` table in the DB
does not have a `project_id` column causing the search to fail
(`ActiveRecord::StatementInvalid (PG::UndefinedColumn: ERROR:  column
equipment_assets.project_id does not exist`). The code of this plugin
already contained a hack to *"placate the use of project_id in search
view"* (see app/models/equipment_asset.rb) but this was not enough as
it turns out.